Got buttons to filter courses, added toggled button states, small bug encountered
Read the scripts.js file to understand how its able to filter the courses using the JSON data.

Instead of doing everything in onload, onload only loads the JSON data, methods declared after this call the data that was loaded.

Also added ID's to the buttons. This allows the filter methods to toggle the states of the buttons so that the button whose filter is currently being applied is shown as active.

Note: Known Bug, the first button click does nothing. Not the first click per button, just the first click in general. No matter the button. Any click after the first on any of the buttons will apply the filters.