Refined filter methods
Instead of one filter method per subject it is now one filter method that works for all subjects. It takes in a parameter for the subject's abbreviation and filters based on that.

Going to take a look into doing it using a database call instead of filtering the list of all courses.