Added Id's to the <tr> elements and the <td> elements that contain the buttons. Renamed scripts.js to coursefilter.js, added button functionality
Id's are assigned in the coursefilter methods. The Id's are based off the course's id numbers.

Added a link in index.jsp to link to the new js file.

addcourse.js contains two methods. AddCourse() and RemoveCourse(). At the moment it just changes the color of the row as well as swap the buttons. Add swaps the button to Remove. Remove swaps the button to Add.